move schema out of submission
