+++ test/integration/submission/schema-manager.spec.ts
@@ -7,7 +7,7 @@ import mongoose from 'mongoose';
 import { GenericContainer } from 'testcontainers';
 import { cleanCollection, findInDb } from '../testutils';
 import _ from 'lodash';
-import * as manager from '../../../src/submission/schema/schema-manager';
+import * as manager from '../../../src/schema/schema-manager';
 import { promisify } from 'bluebird';
 import { SchemasDictionary } from '../../../src/lectern-client/schema-entities';
 const ServerMock: any = require('mock-http-server') as any;
